fix(heartbeat): add overlap_key for cross-source event deduplication
The dedupe key was previously derived from event.fingerprint, which is
source-specific (cron job ID, calendar connection/event ID, notification ID).
This meant the same underlying event surfaced by two different sources
(e.g. a meeting visible in both a cron reminder and a Google Calendar
connection) produced distinct dedupe keys and could generate duplicate
notifications.

Add an overlap_key to PendingEvent computed from category + normalized
title + 15-minute time bucket. Use overlap_key instead of fingerprint
when building the final dedupe hash in evaluate_and_dispatch so identical
events from multiple sources collide on the same key.

Also adds compute_overlap_key() helper to utils.rs and two new unit tests
that verify same-bucket/cross-source collision and different-title/time
separation. Fixes @coderabbitai comment on planner/mod.rs:124.
