+++ src/openhuman/heartbeat/planner/mod.rs
@@ -1,87 +1,44 @@
+//! Heartbeat planner — evaluates upcoming events and dispatches proactive
+//! notifications.
+//!
+//! # Module layout
+//!
+//! | File | Responsibility |
+//! |------|----------------|
+//! | `types.rs` | Shared data types (`HeartbeatCategory`, `PendingEvent`, …) |
+//! | `collectors.rs` | Source-specific collectors (cron, calendar, notifications) |
+//! | `plan.rs` | Delivery-window logic (`plan_delivery_for_event`) |
+//! | `persistence.rs` | Durable notification persistence (`persist_heartbeat_alert`) |
+//! | `utils.rs` | Pure helpers (`sanitize_preview`, `stable_key`) |
+//! | `store.rs` | Dedupe store (`mark_sent`, `prune_old`) |
+
+mod collectors;
+mod persistence;
+mod plan;
+mod store;
+mod types;
+mod utils;
+
+pub use types::PlannerRunSummary;
+
 use std::collections::HashSet;
 
 use chrono::{DateTime, Duration, Utc};
-use serde::Serialize;
-use serde_json::json;
-use sha2::{Digest, Sha256};
 
 use crate::core::event_bus::{publish_global, DomainEvent};
-use crate::openhuman::composio::build_composio_client;
 use crate::openhuman::config::Config;
-use crate::openhuman::cron;
 use crate::openhuman::notifications::bus::publish_core_notification;
-use crate::openhuman::notifications::store as notifications_store;
-use crate::openhuman::notifications::types::{
-    CoreNotificationCategory, CoreNotificationEvent, IntegrationNotification, NotificationStatus,
+use crate::openhuman::notifications::types::CoreNotificationEvent;
+
+use collectors::{
+    collect_calendar_meetings, collect_cron_reminders, collect_relevant_notifications,
 };
+use persistence::persist_heartbeat_alert;
+use plan::plan_delivery_for_event;
+use utils::stable_key;
 
-mod store;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum HeartbeatCategory {
-    Meetings,
-    Reminders,
-    Important,
-}
-
-impl HeartbeatCategory {
-    fn as_str(&self) -> &'static str {
-        match self {
-            Self::Meetings => "meetings",
-            Self::Reminders => "reminders",
-            Self::Important => "important",
-        }
-    }
-
-    fn notification_category(&self) -> CoreNotificationCategory {
-        match self {
-            Self::Meetings => CoreNotificationCategory::Meetings,
-            Self::Reminders => CoreNotificationCategory::Reminders,
-            Self::Important => CoreNotificationCategory::Important,
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-struct PendingEvent {
-    category: HeartbeatCategory,
-    source: String,
-    source_event_id: String,
-    fingerprint: String,
-    title: String,
-    body: String,
-    deep_link: Option<String>,
-    anchor_at: DateTime<Utc>,
-}
-
-#[derive(Debug, Clone)]
-struct PlannedDelivery {
-    stage: &'static str,
-    title: String,
-    body: String,
-    proactive_message: String,
-    allow_external: bool,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct PlannerRunSummary {
-    pub source_events: usize,
-    pub deliveries_attempted: usize,
-    pub deliveries_sent: usize,
-    pub deliveries_skipped_dedup: usize,
-}
-
-impl PlannerRunSummary {
-    fn empty() -> Self {
-        Self {
-            source_events: 0,
-            deliveries_attempted: 0,
-            deliveries_sent: 0,
-            deliveries_skipped_dedup: 0,
-        }
-    }
-}
-
+/// Evaluate all configured notification categories and dispatch any events that
+/// fall within their delivery windows and have not already been sent.
 pub async fn evaluate_and_dispatch(config: &Config, now: DateTime<Utc>) -> PlannerRunSummary {
     let mut summary = PlannerRunSummary::empty();
 
@@ -116,10 +73,15 @@ pub async fn evaluate_and_dispatch(config: &Config, now: DateTime<Utc>) -> Plann
             continue;
         };
 
+        // Use `overlap_key` (content-based: category + title + time-bucket) so
+        // that identical underlying events surfaced by multiple sources
+        // (e.g. the same meeting visible in both cron reminders and a calendar
+        // connection) map to the same dedupe key and only one notification is
+        // delivered.
         let dedupe_key = stable_key(&format!(
             "{}|{}|{}",
             event.category.as_str(),
-            event.fingerprint,
+            event.overlap_key,
             plan.stage
         ));
 
@@ -130,6 +92,28 @@ pub async fn evaluate_and_dispatch(config: &Config, now: DateTime<Utc>) -> Plann
         }
 
         summary.deliveries_attempted += 1;
+
+        let id = format!(
+            "heartbeat:{}:{}:{}",
+            event.category.as_str(),
+            plan.stage,
+            &dedupe_key[..12]
+        );
+
+        // Persist the durable notification record BEFORE marking dedupe, so a
+        // failed write doesn't permanently suppress future retries.
+        if let Err(error) = persist_heartbeat_alert(config, &event, &plan, now) {
+            tracing::warn!(
+                dedupe_key = %dedupe_key,
+                source = %event.source,
+                source_event_id = %event.source_event_id,
+                category = event.category.as_str(),
+                stage = plan.stage,
+                error = %error,
+                "[heartbeat:planner] failed to persist heartbeat alert; skipping delivery"
+            );
+            continue;
+        }
 
         let inserted = match store::mark_sent(
             config,
@@ -161,14 +145,6 @@ pub async fn evaluate_and_dispatch(config: &Config, now: DateTime<Utc>) -> Plann
             continue;
         }
 
-        let id = format!(
-            "heartbeat:{}:{}:{}",
-            event.category.as_str(),
-            plan.stage,
-            &dedupe_key[..12]
-        );
-
-        persist_heartbeat_alert(config, &event, &plan, now);
         publish_core_notification(CoreNotificationEvent {
             id,
             category: event.category.notification_category(),
@@ -205,481 +181,6 @@ pub async fn evaluate_and_dispatch(config: &Config, now: DateTime<Utc>) -> Plann
     summary
 }
 
-fn collect_cron_reminders(config: &Config, now: DateTime<Utc>) -> Vec<PendingEvent> {
-    let lookahead = Duration::minutes(i64::from(
-        config.heartbeat.reminder_lookahead_minutes.max(1),
-    ));
-
-    let jobs = match cron::list_jobs(config) {
-        Ok(jobs) => jobs,
-        Err(error) => {
-            tracing::warn!(error = %error, "[heartbeat:planner] cron list_jobs failed");
-            return Vec::new();
-        }
-    };
-
-    jobs.into_iter()
-        .filter(|job| job.enabled)
-        .filter(|job| is_reminder_like_job(job))
-        .filter(|job| {
-            let delta = job.next_run.signed_duration_since(now);
-            delta <= lookahead && delta >= Duration::minutes(-2)
-        })
-        .map(|job| {
-            let title = job
-                .name
-                .clone()
-                .filter(|name| !name.trim().is_empty())
-                .unwrap_or_else(|| "Reminder".to_string());
-            let fingerprint = stable_key(&format!("cron:{}:{}", job.id, job.next_run.to_rfc3339()));
-            let body = format!(
-                "{} is scheduled at {}.",
-                title,
-                job.next_run.format("%H:%M")
-            );
-
-            PendingEvent {
-                category: HeartbeatCategory::Reminders,
-                source: "cron".to_string(),
-                source_event_id: job.id,
-                fingerprint,
-                title,
-                body,
-                deep_link: Some("/settings/cron-jobs".to_string()),
-                anchor_at: job.next_run,
-            }
-        })
-        .collect()
-}
-
-fn is_reminder_like_job(job: &cron::CronJob) -> bool {
-    if job.delivery.mode.eq_ignore_ascii_case("proactive") {
-        return true;
-    }
-
-    let mut haystack = String::new();
-    if let Some(name) = &job.name {
-        haystack.push_str(name);
-        haystack.push(' ');
-    }
-    if let Some(prompt) = &job.prompt {
-        haystack.push_str(prompt);
-        haystack.push(' ');
-    }
-    haystack.push_str(&job.command);
-
-    let lowered = haystack.to_ascii_lowercase();
-    lowered.contains("remind")
-        || lowered.contains("meeting")
-        || lowered.contains("standup")
-        || lowered.contains("follow up")
-}
-
-async fn collect_calendar_meetings(config: &Config, now: DateTime<Utc>) -> Vec<PendingEvent> {
-    let Some(client) = build_composio_client(config) else {
-        return Vec::new();
-    };
-
-    let connections = match client.list_connections().await {
-        Ok(resp) => resp.connections,
-        Err(error) => {
-            tracing::warn!(error = %error, "[heartbeat:planner] composio list_connections failed");
-            return Vec::new();
-        }
-    };
-
-    let lookahead = Duration::minutes(i64::from(config.heartbeat.meeting_lookahead_minutes.max(1)));
-    let end_window = now + lookahead;
-
-    let mut out = Vec::new();
-    for conn in connections.into_iter().filter(|c| c.is_active()) {
-        let toolkit = conn.normalized_toolkit();
-        if toolkit != "googlecalendar" && toolkit != "google_calendar" && toolkit != "calendar" {
-            continue;
-        }
-
-        let arguments = json!({
-            "connectionId": conn.id,
-            "timeMin": now.to_rfc3339(),
-            "timeMax": end_window.to_rfc3339(),
-            "maxResults": 20
-        });
-
-        let resp = match client
-            .execute_tool("GOOGLECALENDAR_EVENTS_LIST", Some(arguments))
-            .await
-        {
-            Ok(resp) => resp,
-            Err(error) => {
-                tracing::warn!(
-                    toolkit = %toolkit,
-                    connection_id = %conn.id,
-                    error = %error,
-                    "[heartbeat:planner] GOOGLECALENDAR_EVENTS_LIST failed"
-                );
-                continue;
-            }
-        };
-
-        out.extend(extract_calendar_events(
-            &resp.data, &toolkit, &conn.id, now, end_window,
-        ));
-    }
-
-    out
-}
-
-fn extract_calendar_events(
-    value: &serde_json::Value,
-    toolkit: &str,
-    connection_id: &str,
-    start_window: DateTime<Utc>,
-    end_window: DateTime<Utc>,
-) -> Vec<PendingEvent> {
-    let mut out = Vec::new();
-    collect_calendar_events_recursive(
-        value,
-        toolkit,
-        connection_id,
-        start_window,
-        end_window,
-        &mut out,
-    );
-    out
-}
-
-fn collect_calendar_events_recursive(
-    value: &serde_json::Value,
-    toolkit: &str,
-    connection_id: &str,
-    start_window: DateTime<Utc>,
-    end_window: DateTime<Utc>,
-    out: &mut Vec<PendingEvent>,
-) {
-    match value {
-        serde_json::Value::Array(items) => {
-            for item in items {
-                collect_calendar_events_recursive(
-                    item,
-                    toolkit,
-                    connection_id,
-                    start_window,
-                    end_window,
-                    out,
-                );
-            }
-        }
-        serde_json::Value::Object(map) => {
-            if let Some(starts_at) = extract_datetime_from_map(map) {
-                if starts_at >= start_window && starts_at <= end_window {
-                    let title = extract_title_from_map(map);
-                    let source_event_id = map
-                        .get("id")
-                        .and_then(serde_json::Value::as_str)
-                        .or_else(|| map.get("eventId").and_then(serde_json::Value::as_str))
-                        .or_else(|| map.get("icalUID").and_then(serde_json::Value::as_str))
-                        .unwrap_or("calendar-event")
-                        .to_string();
-                    let deep_link = map
-                        .get("htmlLink")
-                        .and_then(serde_json::Value::as_str)
-                        .or_else(|| map.get("hangoutLink").and_then(serde_json::Value::as_str))
-                        .map(ToString::to_string);
-
-                    let fingerprint = stable_key(&format!(
-                        "{}:{}:{}:{}",
-                        toolkit,
-                        connection_id,
-                        source_event_id,
-                        starts_at.to_rfc3339()
-                    ));
-
-                    out.push(PendingEvent {
-                        category: HeartbeatCategory::Meetings,
-                        source: format!("calendar:{toolkit}"),
-                        source_event_id,
-                        fingerprint,
-                        title: title.clone(),
-                        body: format!("{} starts at {}.", title, starts_at.format("%H:%M")),
-                        deep_link,
-                        anchor_at: starts_at,
-                    });
-                }
-            }
-
-            for child in map.values() {
-                collect_calendar_events_recursive(
-                    child,
-                    toolkit,
-                    connection_id,
-                    start_window,
-                    end_window,
-                    out,
-                );
-            }
-        }
-        _ => {}
-    }
-}
-
-fn extract_datetime_from_map(
-    map: &serde_json::Map<String, serde_json::Value>,
-) -> Option<DateTime<Utc>> {
-    let start = map.get("start").and_then(|start| match start {
-        serde_json::Value::Object(start_map) => start_map
-            .get("dateTime")
-            .and_then(serde_json::Value::as_str)
-            .or_else(|| start_map.get("date").and_then(serde_json::Value::as_str)),
-        serde_json::Value::String(s) => Some(s.as_str()),
-        _ => None,
-    });
-
-    let direct = start
-        .or_else(|| map.get("start_time").and_then(serde_json::Value::as_str))
-        .or_else(|| map.get("startTime").and_then(serde_json::Value::as_str))
-        .or_else(|| map.get("starts_at").and_then(serde_json::Value::as_str))
-        .or_else(|| map.get("startsAt").and_then(serde_json::Value::as_str));
-
-    direct.and_then(parse_datetime)
-}
-
-fn extract_title_from_map(map: &serde_json::Map<String, serde_json::Value>) -> String {
-    map.get("summary")
-        .and_then(serde_json::Value::as_str)
-        .or_else(|| map.get("title").and_then(serde_json::Value::as_str))
-        .or_else(|| map.get("name").and_then(serde_json::Value::as_str))
-        .map(|raw| sanitize_preview(raw, 80))
-        .filter(|title| !title.is_empty())
-        .unwrap_or_else(|| "Upcoming meeting".to_string())
-}
-
-fn parse_datetime(raw: &str) -> Option<DateTime<Utc>> {
-    chrono::DateTime::parse_from_rfc3339(raw)
-        .map(|dt| dt.with_timezone(&Utc))
-        .or_else(|_| {
-            chrono::NaiveDate::parse_from_str(raw, "%Y-%m-%d").map(|date| {
-                chrono::DateTime::<Utc>::from_naive_utc_and_offset(
-                    date.and_hms_opt(0, 0, 0).unwrap(),
-                    Utc,
-                )
-            })
-        })
-        .ok()
-}
-
-fn collect_relevant_notifications(config: &Config, now: DateTime<Utc>) -> Vec<PendingEvent> {
-    let items = match notifications_store::list(config, 100, 0, None, Some(0.8)) {
-        Ok(items) => items,
-        Err(error) => {
-            tracing::warn!(error = %error, "[heartbeat:planner] notifications list failed");
-            return Vec::new();
-        }
-    };
-
-    items
-        .into_iter()
-        // Never re-escalate notifications we generated ourselves — that creates a
-        // feedback loop where each heartbeat tick spawns a new "Important event"
-        // with a fresh ID that bypasses the dedupe store.
-        .filter(|item| item.provider != "heartbeat")
-        .filter(|item| {
-            item.status == crate::openhuman::notifications::types::NotificationStatus::Unread
-        })
-        .filter(|item| {
-            item.triage_action
-                .as_deref()
-                .map(|action| action == "escalate" || action == "react")
-                .unwrap_or(false)
-                || item
-                    .raw_payload
-                    .get("urgent")
-                    .and_then(serde_json::Value::as_bool)
-                    .unwrap_or(false)
-        })
-        .filter(|item| now.signed_duration_since(item.received_at) <= Duration::minutes(30))
-        .map(|item| {
-            let title = format!("Important event from {}", item.provider);
-            let body = sanitize_preview(&item.title, 100);
-
-            PendingEvent {
-                category: HeartbeatCategory::Important,
-                source: format!("notification:{}", item.provider),
-                source_event_id: item.id.clone(),
-                fingerprint: stable_key(&format!("notification:{}", item.id)),
-                title,
-                body,
-                deep_link: Some("/notifications".to_string()),
-                anchor_at: item.received_at,
-            }
-        })
-        .collect()
-}
-
-fn plan_delivery_for_event(
-    event: &PendingEvent,
-    config: &Config,
-    now: DateTime<Utc>,
-) -> Option<PlannedDelivery> {
-    let until = event.anchor_at.signed_duration_since(now);
-    let until_minutes = until.num_minutes();
-
-    match event.category {
-        HeartbeatCategory::Meetings => {
-            let lookahead = i64::from(config.heartbeat.meeting_lookahead_minutes.max(1));
-            if until_minutes > 10 && until_minutes <= lookahead {
-                let mins = until_minutes.max(1);
-                return Some(PlannedDelivery {
-                    stage: "heads_up",
-                    title: format!("Meeting soon: {}", event.title),
-                    body: format!("Starts in about {mins} minutes."),
-                    proactive_message: format!(
-                        "You have a meeting coming up in about {mins} minutes: {}.",
-                        event.title
-                    ),
-                    allow_external: false,
-                });
-            }
-            if until_minutes > 0 && until_minutes <= 10 {
-                let mins = until_minutes.max(1);
-                return Some(PlannedDelivery {
-                    stage: "final_call",
-                    title: format!("Upcoming meeting: {}", event.title),
-                    body: format!("Starts in about {mins} minutes."),
-                    proactive_message: format!(
-                        "Your meeting starts in about {mins} minutes: {}.",
-                        event.title
-                    ),
-                    allow_external: true,
-                });
-            }
-            // Wider grace window: heartbeat runs every few minutes, so
-            // tiny post-start windows can miss real meetings.
-            if until_minutes <= 0 && until_minutes >= -10 {
-                return Some(PlannedDelivery {
-                    stage: "starting_now",
-                    title: format!("Meeting starting now: {}", event.title),
-                    body: "This meeting should be starting now.".to_string(),
-                    proactive_message: format!("Your meeting is starting now: {}.", event.title),
-                    allow_external: true,
-                });
-            }
-            None
-        }
-        HeartbeatCategory::Reminders => {
-            let lookahead = i64::from(config.heartbeat.reminder_lookahead_minutes.max(1));
-            if until_minutes > 0 && until_minutes <= lookahead {
-                let mins = until_minutes.max(1);
-                return Some(PlannedDelivery {
-                    stage: "soon",
-                    title: format!("Reminder soon: {}", event.title),
-                    body: format!("Scheduled in about {mins} minutes."),
-                    proactive_message: format!(
-                        "Reminder coming up in about {mins} minutes: {}.",
-                        event.title
-                    ),
-                    allow_external: false,
-                });
-            }
-            // Wider grace window for reminder due state to prevent misses
-            // from tick alignment.
-            if until_minutes <= 0 && until_minutes >= -10 {
-                return Some(PlannedDelivery {
-                    stage: "due",
-                    title: format!("Reminder due: {}", event.title),
-                    body: "A scheduled reminder is due now.".to_string(),
-                    proactive_message: format!("Reminder due now: {}.", event.title),
-                    allow_external: true,
-                });
-            }
-            None
-        }
-        HeartbeatCategory::Important => {
-            if now.signed_duration_since(event.anchor_at) <= Duration::minutes(10) {
-                return Some(PlannedDelivery {
-                    stage: "important_now",
-                    title: event.title.clone(),
-                    body: if event.body.is_empty() {
-                        "A time-sensitive event needs your attention.".to_string()
-                    } else {
-                        event.body.clone()
-                    },
-                    proactive_message: if event.body.is_empty() {
-                        "A time-sensitive event needs your attention.".to_string()
-                    } else {
-                        event.body.clone()
-                    },
-                    allow_external: true,
-                });
-            }
-            None
-        }
-    }
-}
-
-fn sanitize_preview(raw: &str, max_chars: usize) -> String {
-    let clean = raw.split_whitespace().collect::<Vec<_>>().join(" ");
-    if clean.chars().count() <= max_chars {
-        return clean;
-    }
-    let mut trimmed: String = clean.chars().take(max_chars.saturating_sub(1)).collect();
-    trimmed.push('…');
-    trimmed
-}
-
-fn stable_key(seed: &str) -> String {
-    let mut hasher = Sha256::new();
-    hasher.update(seed.as_bytes());
-    hex::encode(hasher.finalize())
-}
-
-fn persist_heartbeat_alert(
-    config: &Config,
-    event: &PendingEvent,
-    plan: &PlannedDelivery,
-    now: DateTime<Utc>,
-) {
-    let notification = IntegrationNotification {
-        id: format!(
-            "heartbeat:{}:{}:{}",
-            event.category.as_str(),
-            plan.stage,
-            &event.fingerprint[..12]
-        ),
-        provider: "heartbeat".to_string(),
-        account_id: Some(event.source_event_id.clone()),
-        title: sanitize_preview(&plan.title, 100),
-        body: sanitize_preview(&plan.body, 180),
-        raw_payload: serde_json::json!({
-            "source": event.source,
-            "category": event.category.as_str(),
-            "stage": plan.stage,
-            "anchor_at": event.anchor_at.to_rfc3339(),
-            "deep_link": event.deep_link.clone(),
-        }),
-        importance_score: Some(match event.category {
-            HeartbeatCategory::Meetings => 0.8,
-            HeartbeatCategory::Reminders => 0.7,
-            HeartbeatCategory::Important => 0.9,
-        }),
-        triage_action: Some("react".to_string()),
-        triage_reason: Some("heartbeat proactive event".to_string()),
-        status: NotificationStatus::Unread,
-        received_at: now,
-        scored_at: Some(now),
-    };
-
-    if let Err(error) = notifications_store::insert_if_not_recent(config, &notification) {
-        tracing::warn!(
-            source = %event.source,
-            source_event_id = %event.source_event_id,
-            category = event.category.as_str(),
-            stage = plan.stage,
-            error = %error,
-            "[heartbeat:planner] failed to persist heartbeat alert"
-        );
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -687,7 +188,13 @@ mod tests {
     use crate::openhuman::cron::{self, Schedule};
     use crate::openhuman::notifications::subscribe_core_notifications;
     use chrono::TimeZone;
+    use serde_json::json;
     use tempfile::TempDir;
+
+    use collectors::extract_calendar_events;
+    use plan::plan_delivery_for_event;
+    use types::{HeartbeatCategory, PendingEvent};
+    use utils::{compute_overlap_key, sanitize_preview};
 
     #[test]
     fn extract_calendar_events_reads_nested_payload() {
@@ -722,6 +229,34 @@ mod tests {
     }
 
     #[test]
+    fn all_day_calendar_events_are_skipped() {
+        let now = Utc.with_ymd_and_hms(2026, 5, 8, 0, 0, 0).unwrap();
+        let payload = json!({
+            "items": [
+                {
+                    "id": "all-day-1",
+                    "summary": "Birthday",
+                    "start": { "date": "2026-05-08" }
+                }
+            ]
+        });
+
+        let events = extract_calendar_events(
+            &payload,
+            "googlecalendar",
+            "conn-1",
+            now,
+            now + Duration::hours(24),
+        );
+
+        assert_eq!(
+            events.len(),
+            0,
+            "all-day events should not be promoted to meetings"
+        );
+    }
+
+    #[test]
     fn reminder_stage_prioritizes_due_window() {
         let mut config = Config::default();
         config.heartbeat.reminder_lookahead_minutes = 15;
@@ -731,6 +266,7 @@ mod tests {
             source: "cron".to_string(),
             source_event_id: "job-1".to_string(),
             fingerprint: "fp-1".to_string(),
+            overlap_key: compute_overlap_key(HeartbeatCategory::Reminders, "Pay rent", now),
             title: "Pay rent".to_string(),
             body: String::new(),
             deep_link: None,
@@ -752,6 +288,11 @@ mod tests {
             source: "calendar:googlecalendar".to_string(),
             source_event_id: "evt-1".to_string(),
             fingerprint: "fp-1".to_string(),
+            overlap_key: compute_overlap_key(
+                HeartbeatCategory::Meetings,
+                "Planning",
+                now + Duration::minutes(45),
+            ),
             title: "Planning".to_string(),
             body: String::new(),
             deep_link: None,
@@ -844,5 +385,51 @@ mod tests {
             summary.deliveries_sent, 0,
             "heartbeat provider notifications must not be re-escalated as Important events"
         );
+    }
+
+    #[test]
+    fn overlap_key_same_for_cross_source_same_event() {
+        // Two different sources that surface the same meeting at the same time
+        // (within the 15-minute bucket) must produce the same overlap_key so
+        // only one notification is dispatched.
+        let anchor = Utc.with_ymd_and_hms(2026, 5, 8, 10, 0, 0).unwrap();
+
+        let key_from_calendar = compute_overlap_key(
+            HeartbeatCategory::Meetings,
+            "Team Standup",
+            anchor,
+        );
+        // A cron job with the same title and an anchor 2 minutes later (same
+        // 15-minute bucket) — different source, same underlying event.
+        let key_from_cron = compute_overlap_key(
+            HeartbeatCategory::Meetings,
+            "Team Standup",
+            anchor + Duration::minutes(2),
+        );
+
+        assert_eq!(
+            key_from_calendar, key_from_cron,
+            "cross-source events in the same 15-min bucket must share an overlap_key"
+        );
+    }
+
+    #[test]
+    fn overlap_key_differs_for_different_titles_or_times() {
+        let anchor = Utc.with_ymd_and_hms(2026, 5, 8, 10, 0, 0).unwrap();
+
+        // Different title → different key.
+        let key_a = compute_overlap_key(HeartbeatCategory::Meetings, "Team Standup", anchor);
+        let key_b = compute_overlap_key(HeartbeatCategory::Meetings, "1:1 With Manager", anchor);
+        assert_ne!(key_a, key_b, "different titles must produce different overlap keys");
+
+        // Same title but more than one bucket apart (>= 15 min) → different key.
+        let key_c =
+            compute_overlap_key(HeartbeatCategory::Meetings, "Team Standup", anchor + Duration::minutes(20));
+        assert_ne!(key_a, key_c, "events in different time buckets must produce different overlap keys");
+
+        // Different category → different key even with same title and time.
+        let key_d =
+            compute_overlap_key(HeartbeatCategory::Reminders, "Team Standup", anchor);
+        assert_ne!(key_a, key_d, "different categories must produce different overlap keys");
     }
 }
